tools/libxenguest: handle more than 16T in precopy_stats
total_written tracks the number of transferred dirty pages.

Acked-by: Wei Liu <wl@xen.org>

+++ tools/include/xenguest.h
@@ -438,7 +438,7 @@ struct xenevtchn_handle;
 struct precopy_stats
 {
     unsigned int iteration;
-    unsigned int total_written;
+    unsigned long total_written;
     long dirty_count; /* -1 if unknown */
 };
